Add check, if base64 is empty

+++ backend/tools/image_validator.py
@@ -5,11 +5,13 @@ import base64
 
 
 async def is_image(str_base64):
-    binary = base64.b64decode(str_base64[str_base64.index(",") + 1:])
-    path = f"/tmp/{uuid.uuid1()}"
     flag = False
+    if str_base64 == []:
+        return flag
+    cuted_base64 = base64.b64decode(str_base64[str_base64.index(",") + 1:])
+    path = f"/tmp/{uuid.uuid1()}"
     with open(path, "wb") as tmpfile:
-        tmpfile.write(binary)
+        tmpfile.write(cuted_base64)
         if imghdr.what(path):
             flag = True
     os.remove(path)
